adding 3 screenshots of each of the years ticker
what has been done here is the original file has been coded to where all the following conditions have been met. conditional formatting met. The ticker symbol.


Yearly change from opening price at the beginning of a given year to the closing price at the end of that year.


The percent change from opening price at the beginning of a given year to the closing price at the end of that year.


The total stock volume of the stock.